pyproject.toml
[tool.poetry]
name = "basicbank"
version = "0.1.0"
description = "Basic banking api for take home project"
authors = ["Mohammed Ali <flak@verify4.com>"]
packages = [{include = "app"}]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.8.2"
pytest = "^8.3.2"
hypothesis = "^6.108.10"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.5"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
SQLAlchemy = "^2.0.32"
pytest-xdist = "^3.6.1"


[tool.pytest.ini_options]
pythonpath = ["."]
filterwarnings = ["ignore:datetime.datetime.utcnow\\(\\) is deprecated:DeprecationWarning"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"